Pull CI status into its own component
