+++ app/src/ui/branches/pull-request-list-item.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import * as moment from 'moment'
 import { Octicon, OcticonSymbol } from '../octicons'
 import { APIRefState } from '../../lib/api'
-import { assertNever } from '../../lib/fatal-error'
+import { CIStatus } from './ci-status'
 
 interface IPullRequestListItemProps {
   /** The title. */
@@ -30,7 +30,6 @@ export class PullRequestListItem extends React.Component<
     const timeAgo = moment(this.props.created).fromNow()
     const { title, author, status } = this.props
     const subtitle = `#${this.props.number} opened ${timeAgo} by ${author}`
-    const ciTitle = `Commit status: ${status}`
     return (
       <div className="pull-request-item">
         <Octicon className="icon" symbol={OcticonSymbol.gitPullRequest} />
@@ -44,25 +43,8 @@ export class PullRequestListItem extends React.Component<
           </div>
         </div>
 
-        <Octicon
-          className={`status status-${status}`}
-          symbol={getSymbolForStatus(status)}
-          title={ciTitle}
-        />
+        <CIStatus status={status} />
       </div>
     )
   }
-}
-
-function getSymbolForStatus(status: APIRefState): OcticonSymbol {
-  switch (status) {
-    case 'pending':
-      return OcticonSymbol.primitiveDot
-    case 'failure':
-      return OcticonSymbol.x
-    case 'success':
-      return OcticonSymbol.check
-  }
-
-  return assertNever(status, `Unknown status: ${status}`)
 }
