+++ app/styles/_ui.scss
@@ -57,3 +57,4 @@
 @import "ui/ref";
 @import "ui/monospaced";
 @import 'ui/initialize-lfs';
+@import 'ui/ci-status';
