+++ app/styles/ui/_branches.scss
@@ -83,23 +83,6 @@
         color: var(--text-secondary-color);
       }
     }
-
-    .status {
-      width: 16px;
-      flex-shrink: 0;
-    }
-
-    .status-pending {
-      color: $orange-700;
-    }
-
-    .status-failure {
-      color: $red-600;
-    }
-
-    .status-success {
-      color: $green-500;
-    }
   }
 
   .pull-request-loading-item {
