+++ app/styles/ui/ci-status.scss
@@ -1,0 +1,16 @@
+.ci-status {
+  width: 16px;
+  flex-shrink: 0;
+}
+
+.ci-status-pending {
+  color: $orange-700;
+}
+
+.ci-status-failure {
+  color: $red-600;
+}
+
+.ci-status-success {
+  color: $green-500;
+}
